Ian Doerr February 17th, 2010
Spent some time updating the code to be compatible with Visual Studio 2008 - mainly replacing older c-functions with their more secure counterparts like strcpy with strcpy_s.

Commented out a few delete statements because they were deleting strings that were stored in the source code, and not in the stack.

Also created .bat files that work with Debug mode and enabled the debug heap when in debug mode.

git-svn-id: file:///e/SVN/3d_engine@69 463a7de7-ff88-454d-8a36-9b7cb3bdd095